@@ -3,6 +3,7 @@ import { AngularFireDatabase } from '@angular/fire/database';
 import { Observable } from 'rxjs';
 import { TransferState, makeStateKey } from '@angular/platform-browser';
 import { startWith, tap } from 'rxjs/operators';
+import { trace } from '@angular/fire/performance';
 
 @Component({
   selector: 'app-database',
@@ -22,7 +23,10 @@ export class DatabaseComponent implements OnInit {
     const doc = database.object('test');
     const key = makeStateKey(doc.query.toString());
     const existing = state.get(key, undefined);
-    this.testObjectValue$ = doc.valueChanges().pipe(existing ? startWith(existing) : tap(it => state.set(key, it)));
+    this.testObjectValue$ = doc.valueChanges().pipe(
+      trace('database'),
+      existing ? startWith(existing) : tap(it => state.set(key, it))
+    );
   }
 
   ngOnInit(): void {
