@@ -3,6 +3,7 @@ import { AngularFireStorage } from '@angular/fire/storage';
 import { Observable, of } from 'rxjs';
 import { startWith, tap } from 'rxjs/operators';
 import { makeStateKey, TransferState } from '@angular/platform-browser';
+import { trace } from '@angular/fire/performance';
 
 const TRANSPARENT_PNG = 'data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAQAAAC1HAwCAAAAC0lEQVR42mNkYAAAAAYAAjCB0C8AAAAASUVORK5CYII='
 
@@ -24,7 +25,11 @@ export class StorageComponent implements OnInit {
     const icon = storage.ref('google-g.png');
     const key = makeStateKey('google-icon-url');
     const existing = state.get(key, undefined);
-    this.downloadUrl$ = existing ? of(existing) : icon.getDownloadURL().pipe(tap(it => state.set(key, it)), startWith(TRANSPARENT_PNG));
+    this.downloadUrl$ = existing ? of(existing) : icon.getDownloadURL().pipe(
+      trace('storage'),
+      tap(it => state.set(key, it)),
+      startWith(TRANSPARENT_PNG)
+    );
   }
 
   ngOnInit(): void {
