@@ -1,11 +1,16 @@
-import { NgModule } from '@angular/core';
+import { NgModule, Optional } from '@angular/core';
 import { AngularFirePerformance } from './performance';
+import { PerformanceMonitoringService } from './performance.service';
 
 @NgModule({
   providers: [ AngularFirePerformance ]
 })
 export class AngularFirePerformanceModule {
-  constructor(_: AngularFirePerformance) {
-    // DI inject AFP here for the automatic data collection
-  } 
+  constructor(
+    perf: AngularFirePerformance,
+    @Optional() _: PerformanceMonitoringService
+  ) {
+    // call anything here to get perf loading
+    perf.dataCollectionEnabled
+  }
 }
