@@ -6,6 +6,7 @@ import { BuildTarget, FirebaseTools, FSHost } from "../interfaces";
 import { writeFileSync, renameSync, existsSync, readFileSync } from "fs";
 import { copySync, removeSync } from "fs-extra";
 import { join, dirname } from "path";
+import { execSync } from "child_process";
 import {
   defaultFunction,
   defaultPackage,
@@ -15,6 +16,8 @@ import { experimental } from "@angular-devkit/core";
 import { SchematicsException } from "@angular-devkit/schematics";
 import { satisfies } from "semver";
 
+const escapeRegExp = (str:String) => str.replace(/[\-\[\]\/\{\}\(\)\*\+\?\.\\\^\$\|]/g, "\\$&");
+
 const moveSync = (src: string, dest: string) => {
   copySync(src, dest);
   removeSync(src);
@@ -23,13 +26,44 @@ const moveSync = (src: string, dest: string) => {
 const deployToHosting = (
   firebaseTools: FirebaseTools,
   context: BuilderContext,
-  workspaceRoot: string
+  workspaceRoot: string,
+  preview: boolean
 ) => {
-  return firebaseTools.deploy({
-    // tslint:disable-next-line:no-non-null-assertion
-    only: "hosting:" + context.target!.project,
-    cwd: workspaceRoot
-  });
+
+  if (preview) {
+    const port = 5000; // TODO make this configurable
+
+    setTimeout(() => {
+      execSync(`open http://localhost:${port} || true`);
+    }, 1500);
+
+    return firebaseTools.serve({ port, targets: ["hosting"]}).then(() =>
+      require('inquirer').prompt({
+        type: 'confirm',
+        name: 'deployProject',
+        message: "Would you like to deploy your application to Firebase Hosting?"
+      })
+    ).then(({ deployProject }: { deployProject: boolean }) => {
+      if (deployProject) {
+        return firebaseTools.deploy({
+          // tslint:disable-next-line:no-non-null-assertion
+          only: "hosting:" + context.target!.project,
+          cwd: workspaceRoot
+        });
+      } else {
+        return Promise.resolve();
+      }
+    });
+
+  } else {
+
+    return firebaseTools.deploy({
+      // tslint:disable-next-line:no-non-null-assertion
+      only: "hosting:" + context.target!.project,
+      cwd: workspaceRoot
+    });
+
+  }
 };
 
 const defaultFsHost: FSHost = {
@@ -40,21 +74,44 @@ const defaultFsHost: FSHost = {
 
 const getVersionRange = (v: number) => `^${v}.0.0`;
 
-const getPackageJson = (workspaceRoot: string) => {
-  const versions = {
+const getPackageJson = (context: BuilderContext, workspaceRoot: string) => {
+  const dependencies = {
     'firebase-admin': 'latest',
-    'firebase-functions': 'latest',
+    'firebase-functions': 'latest'
+  };
+  const devDependencies = {
     'firebase-functions-test': 'latest'
   };
-  if (existsSync(join(workspaceRoot, 'package.json'))) {
-    try {
-      const content = JSON.parse(readFileSync(join(workspaceRoot, 'package.json')).toString());
-      Object.keys(versions).forEach((p: string) => {
-        versions[p] = content.devDependencies[p] || content.dependencies[p] || versions[p];
+  const npmList = execSync('npm ls || true').toString();
+  Object.keys(dependencies).forEach((dependency: string) => {
+    const npmLsMatch = npmList.match(` ${escapeRegExp(dependency)}@.+\\w`);
+    if (npmLsMatch) { dependencies[dependency] = npmLsMatch[0].split(`${dependency}@`)[1] }
+  });
+  Object.keys(devDependencies).forEach((devDependency: string) => {
+    const npmLsMatch = npmList.match(` ${escapeRegExp(devDependency)}@.+\\w`);
+    if (npmLsMatch) { devDependencies[devDependency] = npmLsMatch[0].split(`${devDependency}@`)[1] }
+  });
+  if (existsSync(join(workspaceRoot, 'angular.json'))) {
+    const angularJson = JSON.parse(readFileSync(join(workspaceRoot, 'angular.json')).toString());
+    const server = angularJson.projects[context.target!.project].architect.server;
+    const serverOptions = server && server.options;
+    const externalDependencies = serverOptions && serverOptions.externalDependencies || [];
+    const bundleDependencies = serverOptions && serverOptions.bundleDependencies;
+    if (bundleDependencies == false) {
+      if (existsSync(join(workspaceRoot, 'package.json'))) {
+        const packageJson = JSON.parse(readFileSync(join(workspaceRoot, 'package.json')).toString());
+        Object.keys(packageJson.dependencies).forEach((dependency: string) => {
+          dependencies[dependency] = packageJson.dependencies[dependency];
+        });
+      } // TODO should we throw?
+    } else {
+      externalDependencies.forEach(externalDependency => {
+        const npmLsMatch = npmList.match(` ${escapeRegExp(externalDependency)}@.+\\w`);
+        if (npmLsMatch) { dependencies[externalDependency] = npmLsMatch[0].split(`${externalDependency}@`)[1] }
       });
-    } catch {}
-  }
-  return defaultPackage(versions["firebase-admin"], versions["firebase-functions"], versions["firebase-functions-test"]);
+    }
+  } // TODO should we throw?
+  return defaultPackage(dependencies, devDependencies);
 };
 
 export const deployToFunction = async (
@@ -107,8 +164,9 @@ export const deployToFunction = async (
 
   fsHost.writeFileSync(
     join(dirname(serverOut), "package.json"),
-    getPackageJson(workspaceRoot)
+    getPackageJson(context, workspaceRoot)
   );
+
   fsHost.writeFileSync(
     join(dirname(serverOut), "index.js"),
     defaultFunction(serverOut)
@@ -119,15 +177,32 @@ export const deployToFunction = async (
     join(newClientPath, "index.original.html")
   );
 
-  context.logger.info("Deploying your Angular Universal application...");
-
   if (preview) {
-    context.logger.info(
-      "Your Universal application is now ready for preview. Use `firebase serve` in the output directory of your workspace to test the setup."
-    );
-    return Promise.resolve();
+    const port = 5000; // TODO make this configurable
+
+    setTimeout(() => {
+      execSync(`open http://localhost:${port} || true`);
+    }, 1500);
+
+    return firebaseTools.serve({ port, targets: ["hosting", "functions"]}).then(() =>
+      require('inquirer').prompt({
+        type: 'confirm',
+        name: 'deployProject',
+        message: "Would you like to deploy your application to Firebase Hosting & Cloud Functions?"
+      })
+    ).then(({ deployProject }: { deployProject: boolean }) => {
+      if (deployProject) {
+        return firebaseTools.deploy({
+          only: `hosting:${context.target!.project},functions:ssr`,
+          cwd: workspaceRoot
+        });
+      } else {
+        return Promise.resolve();
+      }
+    });
   } else {
     return firebaseTools.deploy({
+      only: `hosting:${context.target!.project},functions:ssr`,
       cwd: workspaceRoot
     });
   }
@@ -167,6 +242,20 @@ export default async function deploy(
   try {
     let success: { hosting: string };
 
+    const winston = require('winston');
+    const tripleBeam = require('triple-beam');
+
+    firebaseTools.logger.add(
+      new winston.transports.Console({
+        level: "info",
+        format: winston.format.printf((info) =>
+          [info.message, ...(info[tripleBeam.SPLAT] || [])]
+            .filter((chunk) => typeof chunk == "string")
+            .join(" ")
+        ),
+      })
+    );
+
     if (ssr) {
       success = await deployToFunction(
         firebaseTools,
@@ -179,17 +268,11 @@ export default async function deploy(
       success = await deployToHosting(
         firebaseTools,
         context,
-        context.workspaceRoot
+        context.workspaceRoot,
+        preview
       );
     }
 
-    if (!preview) {
-      context.logger.info(
-        `🚀 Your application is now available at https://${
-          success.hosting.split("/")[1]
-        }.firebaseapp.com/`
-      );
-    }
   } catch (e) {
     context.logger.error(e.message || e);
   }
