@@ -1,9 +1,15 @@
+// TODO allow these to be configured
 export const NodeVersion = 10;
+const FUNCTION_NAME = 'ssr';
+const FUNCTION_REGION = 'us-central1';
+const RUNTIME_OPTIONS = {
+  timeoutSeconds: 60,
+  memory: '1GB'
+}
 
 export const defaultPackage = (
-  firebaseAdminVersion: string,
-  firebaseFunctionsVersion: string,
-  firebaseFunctionsTestVersion: string
+  dependencies: {[key:string]: string},
+  devDependencies: {[key:string]: string}
 ) => `{
   "name": "functions",
   "description": "Angular Universal Application",
@@ -18,13 +24,8 @@ export const defaultPackage = (
   "engines": {
     "node": "${NodeVersion}"
   },
-  "dependencies": {
-    "firebase-admin": "${firebaseAdminVersion}",
-    "firebase-functions": "${firebaseFunctionsVersion}"
-  },
-  "devDependencies": {
-    "firebase-functions-test": "${firebaseFunctionsTestVersion}"
-  },
+  "dependencies": ${JSON.stringify(dependencies, null, 4)},
+  "devDependencies": ${JSON.stringify(devDependencies, null, 4)},
   "private": true
 }
 `;
@@ -33,5 +34,11 @@ export const defaultFunction = (
   path: string
 ) => `const functions = require('firebase-functions');
 
-exports.ssr = functions.https.onRequest(require('./${path}/main').app());
+const expressApp = require('./${path}/main').app();
+
+exports.${FUNCTION_NAME} = functions
+  .region('${FUNCTION_REGION}')
+  .runWith(${JSON.stringify(RUNTIME_OPTIONS)})
+  .https
+  .onRequest(expressApp);
 `;
