@@ -41,6 +41,12 @@ function generateHostingConfig(project: string, dist: string) {
   };
 }
 
+function generateFunctionsConfig(dist: string) {
+  return {
+    source: dirname(dist)
+  };
+}
+
 export function generateFirebaseJson(
   tree: Tree,
   path: string,
@@ -52,6 +58,7 @@ export function generateFirebaseJson(
     ? safeReadJSON(path, tree)
     : emptyFirebaseJson(dirname(serverOutput));
 
+  /* TODO do we want to prompt for override?
   if (
     firebaseJson.hosting &&
     ((Array.isArray(firebaseJson.hosting) &&
@@ -61,7 +68,7 @@ export function generateFirebaseJson(
     throw new SchematicsException(
       `Target ${project} already exists in firebase.json`
     );
-  }
+  }*/
 
   const newConfig = generateHostingConfig(project, dist);
   if (firebaseJson.hosting === undefined) {
@@ -71,6 +78,8 @@ export function generateFirebaseJson(
   } else {
     firebaseJson.hosting = [firebaseJson.hosting!, newConfig];
   }
+
+  firebaseJson.functions = generateFunctionsConfig(dist);
 
   overwriteIfExists(tree, path, stringifyFormatted(firebaseJson));
 }
@@ -116,6 +125,11 @@ export const setupUniversalDeployment = (config: {
 
   const staticOutput = project.architect.build.options.outputPath;
   const serverOutput = project.architect.server.options.outputPath;
+
+  // Add @firebase/firestore to externalDependencies 
+  const externalDependencies = project.architect.server.options.externalDependencies || [];
+  externalDependencies.push('@firebase/firestore');
+  project.architect.server.options.externalDependencies = externalDependencies;
 
   project.architect.deploy = {
     builder: '@angular/fire:deploy',
